Created cpp files and extra functions. Converted data types to float for storage.

+++ src/validate.cpp
@@ -1,9 +1,9 @@
 #include "header.hpp"
 
-bool isValidDeposit(int depoAmt, int currentBalance) {
+bool isValidDeposit(float depoAmt, float currentBalance) {
   return (depoAmt > 0) && ((depoAmt + currentBalance) <= MAX_LIMIT);
 }
 
-bool isValidWithdrawal(int withdrawAmt, int currentBalance) {
+bool isValidWithdrawal(float withdrawAmt, float currentBalance) {
   return (withdrawAmt > 0) && (withdrawAmt <= currentBalance);
 }
